chore: rename namespace and apply style fixes
- Move opening brace to next line on new method signatures.
- Rename src/Retry/Standard -> src/Retry/V3 and tests/Retry/Standard -> tests/Retry/V3
- Update namespaces in moved files
- Rename alias StandardRetryMiddleware -> RetryV3Middleware at all import and call sites
- OptInTest: move ' true' / 'true ' from nonMatchingValues to a new matchingValues
  provider (with extra whitespace variants) so tests match the trim($value) === 'true'
  contract in OptIn::isEnabled
